+++ notes.md
@@ -1,0 +1,14 @@
+if(name === 'level')
+    switch(value) {
+      case '5':
+        setTimer(30);
+        break;
+      case '8':
+        setTimer(45);
+        break;
+      case '10':
+        setTimer(60);
+        break;
+      default:
+        break;
+    }
