+++ src/App.js
@@ -1,18 +1,16 @@
 import './App.css';
 import { Routes, Route } from "react-router-dom";
-// import SelfList from './features/remember/selfList';
-// import SelfAns from './features/remember/selfAns';
-import Home from './components/home/home';
-import Result from './features/remember/result';
-import Play from './components/play/play';
+import Home from './components/home';
+import Result from './components/result';
+import Play from './components/play';
+import List from './components/list';
 
 function App() {
   return (
     <div className="app">
       <Routes>
         <Route exact path="/" element={<Home />} />
-        {/* <Route path="/selfList" element={<SelfList />} />
-        <Route path="/selfAnsList" element={<SelfAns />} /> */}
+        <Route path="list" element={<List />} />
         <Route path="play" element={<Play />} />
         <Route path="result" element={<Result />} />
       </Routes>
